Add Pyodide handles (#49)
ryanking13 ran into the problem in pyodide/pyodide#3184 (comment) that it's
inconvenient to send results from one @run_in_pyodide invocation to another.
This adds a mechanism to do so. A SeleniumHandle (perhaps can be named better)
can be returned to the host and when passed back into another @run_in_pyodide
function one can use it to access the object.

This should allow us to make fixtures specifically for use with @run_in_pyodide
functions that return one or more SeleniumHandles.