ENH Add rename_fixture and standalone decorators
We use a bunch of different selenium variants but for the most part they are
drop-in replacements for each other. For instance, when using `selenium_standalone`,
we always just say `selenium = selenium_standalone` at the top of the function. I
think it's clearer if we use a decorator to rename the fixture.