Better handling for unpickle failure (#160)
When we pickle the object, also store its repr. If we fail to unpickle,
include the repr in the error message.

Also, add the info to the error using `add_node`. This requires Python 3.11
so I bumped the requires_python.

This requires Pytest 8 for the test to work, so I needed to merge #161 first.